onclick_handler callback function now passed DOM element
When creating a wordcloud, make_me_a_venncloud() has an optional
parameter 'onclick_handler' that specifies a callback function for
when the user clicks on the <span> containing a word in the wordcloud.

Previously the 'onclick_handler' function was just passed the text of
the word in the wordcloud.  Now, the 'onclick_handler' function is
passed both the text AND the span element.  This allows the callback
function to modify the <span> element, e.g. by adding/removing CSS
classes.

This change should not affect existing 'onclick_handler' callback
functions that only expect a single argument.