Fixed bug when resizing window on corpus-view page
Added (expected) ID to <nav> element in corpus_wordcloud.tpl

Fixed guard condition (in corpus_wordcloud.tpl and document.tpl)
checking to see if the expected ID actually exists.

WaveformVisualizer instances can be passed a callback function that is
invoked when the browser window or size of the controls changes.  The
callback function for corpus_wordcloud.tpl was copied from
document.tpl.  The callback function assumed that the <nav> element
for the waveform had ID "waveform_visualizer", but the <nav> element
in corpus_wordcloud.tpl did not have a DOM ID.  Because the DOM ID was
missing, and the guard condition wasn't implemented correctly, the
callback function incorrectly computed the padding on the <body> tag,
and the wordcloud would be partially hidden behind the waveform
visualizer after the browser window was resized.