CSS tweak - active token no longer changes color on mouseover
When a token is moused-over, it's background color temporarily changes
to light-gray with a medium-gray border.  The "active" (most recently
clicked on) token in the wordcloud has a pink background.

With this commit, the active token no longer changes color from pink
to light-gray on mouseover, but the medium-gray border still appears
when the active-token is moused-over.