Fixed bug where clicking "Junk" icon raised an error
In junk_this_pseudoterm(), the value of the 'dataset' field in an AJAX
request was being set based on some data attached to the mouse event
object:

    send.dataset = event.data.corpus_name;

but the mouse event object was using the fieldname 'corpus' instead of
'corpus_name'.  The 'dataset' field of the AJAX request was being set
to an empty string, which caused the Python code in vapor_bottle.py
that handles the AJAX request to raise an exception.