Improved font-size scaling for words in wordclouds.
The font size of terms in the wordcloud can now be specified by a
function that takes into account the attributes of all terms in the
wordcloud.

If the maximum attribute level exceeds a threshold, the font size of
words is now scaled using a cube root instead of a square root,
reducing the maximum possible font size.